Remove stale references to kind-sysbox.
This change removes last traces of kind-sysbox in the Sysbox
test suite.

kind-sysbox was an experimental fork of the k8s.io KinD project that was changed
to integrate with Sysbox. It was used in the sysbox test suite to test
k8s-in-docker. However, it has been deprecated and we now use kindbox instead.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>