First commit of customized minikube kicbase image.
This image is needed to run Minikube inside a Sysbox container.

Minikube uses the "gcr.io/k8s-minikube/kicbase" image to deploy K8s-in-Docker.
However, that image requires cgroups v2 be enabled on the host whenever it's
running on a rootless environment (more specifically within a user-namespace).

This requirement stems from the fact that cgroups v2 formalizes the concept of
"cgroup delegation", where the container is allowed to manage a subset of the
cgroup hierarchy.

In Sysbox containers however, this requirement is too stringent, as Sysbox
supports cgroup delegation with cgroups v1 too (i.e., it uses a technique
to ensure that a container can manage a cgroup hierarchy in isolation
whether the host is configured with cgroups v1 or v2).

Therefore, this new image relaxes the cgroup v2 requirement by modifying the
entrypoint script in the gcr.io/k8s-minikube/kicbase image.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>