Update images relying on systemd services to mask certain services
The goal here is to prevent systemd from wasting cycles attempting to run services/units that are not applicable within a sys container. These changes are part of the effort initiated to reduce Sysbox's kernel-module dependencies (i.e., configfs, debugfs, tracefs).

As part of these changes we have also simplified the docker installation process in the Dockerfiles being modified.

Also, the images associated to every Dockerfile being touched, as well as those that depend on these ones, have been rebuilt and updated in ghcr.io.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>